remove Utils package and place functions in appropriate classes
